+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 ## 🧩 Addon Updates (2026-02-10)
 
-**The Quartermaster** — v1.0.16  
+**The Quartermaster** — v1.0.16a  
 
 **Changes:**  
 • Added Discord Feature Request: Clear button on Search Boxes. This only appears when there is something to clear.  
 
 **Fixes:**  
-• None.
+• Updated the Materials expansion-tag lookup to fix issue found with No crafting materials found….
+• Fixed an item count tooltip error on Materials screen
 
 **Known issues:**  
 • None currently known.